Moved log4j configuration to the root of resources otherwise it wouldn't work. Added throne main package into the configuration file of log4j with DEBUG mode on.
Loggers are now added to ThroneAuthenticator, ThroneAuthorizator and EmbeddedServer.

EmbeddedServer is now using no SSL context (not that it was before but it's now configured manually that it won't be using SSL context). But it's using the custom authenticator and authorizator.

Created empty test suites.

Added Eclipse PAHO MQTT client for with testing scope.